@@ -12,30 +12,32 @@
     "report": "vue-cli-service build --report"
   },
   "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.7.2",
     "@mdi/font": "^4.9.95",
-    "@tinymce/tinymce-vue": "^3.2.8",
-    "axios": "^0.19.2",
+    "@tinymce/tinymce-vue": "^6.1.0",
+    "axios": "^1.8.2",
     "chart.js": "^3.4.0",
-    "core-js": "^2.6.12",
+    "core-js": "^3.41.0",
     "css-loader": "^3.6.0",
+    "mitt": "^3.0.1",
+    "sass-loader": "^16.0.5",
     "style-loader": "^1.3.0",
     "stylus": "^0.54.8",
     "stylus-loader": "^3.0.2",
-    "tinymce": "^5.9.2",
-    "vue": "^2.6.12",
-    "vue-advanced-cropper": "^0.15.3",
+    "tinymce": "^7.7.1",
+    "vue": "^3.5.13",
+    "vue-advanced-cropper": "^2.8.9",
     "vue-chartjs": "^3.5.1",
+    "vue-router": "^4.5.0",
     "vuedraggable": "^2.24.3",
-    "vuetify": "^2.5.0"
+    "vuetify": "^3.7.15",
+    "webpack": "^5.98.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.11.0",
-    "@vue/cli-service": "^3.11.0",
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-service": "~5.0.8",
     "sass": "~1.32.12",
-    "sass-loader": "^7.3.1",
-    "vue-cli-plugin-vuetify": "^0.6.3",
-    "vue-template-compiler": "^2.6.12",
-    "vuetify-loader": "^1.7.1"
+    "vue-cli-plugin-vuetify": "~2.5.8"
   },
   "postcss": {
     "plugins": {
