@@ -1,7 +1,8 @@
 var config={
     UrlPrefix:'',
     
-    BackendBase:'http://dev-crm.test/backend',
+    //BackendBase:'http://dev-crm.test/backend',
+    BackendBase:'http://localhost:5000',
     //MessengerWS:'ws://dev-crm.test/backend/messenger/ws',
     MessengerWS:'ws://localhost:5000/messenger/ws',
     MessengerSignal:'/messenger/sms.ogg',
