@@ -59,7 +59,6 @@
 </template>
 
 <script>
-  import { bus } from '../../main'
   import Slide from './1_to_m/slide';
   import one_to_m_form from './1_to_m/form';
   
@@ -87,7 +86,7 @@
   props:['form','field'],
   created(){
     let t=this
-    bus.$on( // обновление значений в 1_to_m
+    this.$bus.$on( // обновление значений в 1_to_m
       '1_to_m:upload_values:'+this.field.name,
       values=>{
         //console.log('opload_values:',values);
@@ -168,7 +167,7 @@
       this.refresh++
     },
     open_new_dialog(){
-      bus.$emit('1_to_m_open_new_dialog:'+this.field.name);
+      this.$bus.$emit('1_to_m_open_new_dialog:'+this.field.name);
     },
     // множественная загрузка файлов
     start_dialog_multiload(){ // диалог для множественной загрузки файлов
