@@ -114,7 +114,6 @@
 </template>
 
 <script>
-import { bus } from '../../main'
 import { field_update,check_fld } from './field_functions'
 export default {
   
@@ -152,7 +151,7 @@ export default {
             this.parent(this.field)
           }
           else{
-            bus.$emit('change_field',this.field);
+            this.$bus.$emit('change_field',this.field);
           }          
         },
         refresh(){ 
@@ -177,7 +176,7 @@ export default {
 
     if(!t.parent){
       
-      bus.$on(`field-update:${t.field.name}`,t._field_update )
+      this.$bus.$on(`field-update:${t.field.name}`,t._field_update )
     }
     else{
       console.log('parent:',parent)
@@ -269,7 +268,7 @@ export default {
         if(this.parent)
             this.parent({value:f.value,error:f.error,name:f.name})
         else
-          bus.$emit('change_field',f)
+          this.$bus.$emit('change_field',f)
       }
     },
     regexp_check(){
