@@ -12,8 +12,8 @@
                                 <a href="" style="text-decoration: none; " @click.prevent="show_tooltip(header.tooltip);"><v-icon style="font-size: 8pt;" small color="primary">fa fa-question</v-icon></a>
                             </span>
                             <div>
-                                <a href="#" @click.prevent="go_sort(idx,'asc')" class="sort_desc" :class=""><v-icon color="green" small>keyboard_arrow_down</v-icon></a>
-                                  <a href="#" @click.prevent="go_sort(idx,'desc')" class="sort_asc" :class=""><v-icon color="red" small>keyboard_arrow_up</v-icon></a>
+                                <a href="#" @click.prevent="go_sort(idx,'asc')" class="sort_desc"><v-icon color="green" small>keyboard_arrow_down</v-icon></a>
+                                  <a href="#" @click.prevent="go_sort(idx,'desc')" class="sort_asc"><v-icon color="red" small>keyboard_arrow_up</v-icon></a>
                             </div>
                         </th>
                     </tr>
