@@ -1,10 +1,18 @@
-import Vue from 'vue';
-import Vuetify from 'vuetify/lib';
+import { createVuetify } from 'vuetify';
+import * as components from 'vuetify/components';
+import * as directives from 'vuetify/directives';
 
-Vue.use(Vuetify);
+//import { aliases, mdi } from 'vuetify/iconsets/mdi-svg'
+import { aliases, fa } from 'vuetify/iconsets/fa4'
 
-export default new Vuetify({
+export default createVuetify({
+  components,
+  directives,
   icons: {
-    iconfont: 'mdi',
+    //defaultSet: 'mdi', // Используем Material Design Icons
+    defaultSet: 'fa',
+    sets: {
+      fa
+    },
   },
 });
